Replace PNG images with SVG generated from Dot
All of the affected images represent FleCSI control models.

A side effect of this commit is to correct/update the image of the
Poisson application's control model.  Previously, the namespace was
"standalone" instead of "poisson"; there was no "solve" control point;
and there existed a cycle in the graph.  The new image was produced
from the current version of the application.

Other small changes include image widths specified in the tutorial
pages being tweaked to normalize the text size and the "standalone"
image being renamed to "poisson".