Update control model base and corresponding documentation.
This updates the control model base to include the control_point and meta_point
aliases and uses util::types (called list following the topology alias) in place
of std::tuple for specifying the control_points graph.  This deprecates the use
of std::tuple for specifying control_points as well as the run::control_point
alias.  Related documentation is updated to reflect these changes.