Limit lifetime of control_base policies to that of execute.
This ties the lifetime of policies inheriting from control_base to the execute
method.  A simple exception type is added to facilitate early termination of
control policy execution and error handling in the control policy constructor.
State used to represent the control model graph is moved into a registry type.